Align Windows tests with shared CI
