+++ scripts/package-manager-command.test.ts
@@ -1,20 +1,19 @@
-import { describe, it } from "node:test";
-import assert from "node:assert/strict";
+import { describe, it, expect } from "vitest";
 import {
   resolveCommandInvocation,
   resolvePnpmCommand,
-} from "../scripts/package-manager-command.mjs";
+} from "./package-manager-command.mjs";
 
 describe("resolvePnpmCommand", () => {
   it("uses pnpm.cmd when launched directly on Windows", () => {
-    assert.deepEqual(resolvePnpmCommand({}, "win32"), {
+    expect(resolvePnpmCommand({}, "win32")).toEqual({
       command: "pnpm.cmd",
       prefixArgs: [],
     });
   });
 
   it("uses pnpm on non-Windows platforms", () => {
-    assert.deepEqual(resolvePnpmCommand({}, "darwin"), {
+    expect(resolvePnpmCommand({}, "darwin")).toEqual({
       command: "pnpm",
       prefixArgs: [],
     });
@@ -26,43 +25,38 @@ describe("resolvePnpmCommand", () => {
       "win32",
     );
 
-    assert.equal(resolved.command, process.execPath);
-    assert.deepEqual(resolved.prefixArgs, ["C:\\Users\\runner\\pnpm.cjs"]);
+    expect(resolved.command).toBe(process.execPath);
+    expect(resolved.prefixArgs).toEqual(["C:\\Users\\runner\\pnpm.cjs"]);
   });
 
   it("wraps Windows command shims with cmd.exe", () => {
-    assert.deepEqual(
+    expect(
       resolveCommandInvocation(
         "pnpm.cmd",
         ["--filter", "@step-cli/core", "run", "build"],
         "win32",
         {},
       ),
-      {
-        command: "cmd.exe",
-        args: ["/d", "/s", "/c", "pnpm.cmd --filter @step-cli/core run build"],
-      },
-    );
-    assert.deepEqual(resolveCommandInvocation("tool.bat", [], "win32", {}), {
+    ).toEqual({
+      command: "cmd.exe",
+      args: ["/d", "/s", "/c", "pnpm.cmd --filter @step-cli/core run build"],
+    });
+    expect(resolveCommandInvocation("tool.bat", [], "win32", {})).toEqual({
       command: "cmd.exe",
       args: ["/d", "/s", "/c", "tool.bat"],
     });
   });
 
   it("does not wrap normal executables", () => {
-    assert.deepEqual(
+    expect(
       resolveCommandInvocation("node.exe", ["--version"], "win32"),
-      {
-        command: "node.exe",
-        args: ["--version"],
-      },
-    );
-    assert.deepEqual(
-      resolveCommandInvocation("pnpm", ["--version"], "darwin"),
-      {
-        command: "pnpm",
-        args: ["--version"],
-      },
-    );
+    ).toEqual({
+      command: "node.exe",
+      args: ["--version"],
+    });
+    expect(resolveCommandInvocation("pnpm", ["--version"], "darwin")).toEqual({
+      command: "pnpm",
+      args: ["--version"],
+    });
   });
 });
